style: info icons on Deck name, Page icon, Toggle mode fields
Extracts a shared FieldHint component (click-to-open popover) and
uses it on all three Deck & structure fields, replacing the always-on
hint paragraphs. LocalCheckbox and FontSizePicker now delegate to
FieldHint instead of owning their own popover state. Checkbox row
gap reduced from --space-4 to --space-3 for a more compact list.
