+++ web/src/components/CardOptionsForm/CardOptionsForm.module.css
@@ -27,6 +27,12 @@
   color: var(--color-text-primary);
 }
 
+.labelRow {
+  display: flex;
+  align-items: center;
+  gap: 0.375rem;
+}
+
 .sectionHint {
   font-size: var(--text-xs);
   color: var(--color-text-secondary);
@@ -128,7 +134,7 @@
 .groupOptions {
   display: flex;
   flex-direction: column;
-  gap: var(--space-4);
+  gap: var(--space-3);
 }
