+++ web/src/components/LocalCheckbox.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useRef, useState } from 'react';
 import styles from '../styles/shared.module.css';
+import { FieldHint } from './FieldHint';
 
 interface Props {
   label: string;
@@ -16,21 +16,7 @@ function LocalCheckbox({
   onChecked,
   badge,
 }: Readonly<Props>) {
-  const [open, setOpen] = useState(false);
-  const wrapperRef = useRef<HTMLDivElement>(null);
   const id = `chk-${label}`;
-
-  useEffect(() => {
-    if (!open) return;
-    function handleOutside(e: MouseEvent) {
-      if (wrapperRef.current && !wrapperRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    }
-    document.addEventListener('mousedown', handleOutside);
-    return () => document.removeEventListener('mousedown', handleOutside);
-  }, [open]);
-
   return (
     <div className={styles.checkboxRow}>
       <div className={styles.checkboxControl}>
@@ -47,28 +33,7 @@ function LocalCheckbox({
       </div>
       <div className={styles.checkboxRight}>
         {badge && <span className={styles.checkboxBadge}>{badge}</span>}
-        {description && (
-          <div ref={wrapperRef} className={styles.infoAnchor}>
-            <button
-              type="button"
-              className={styles.infoButton}
-              aria-label={open ? 'Hide description' : 'Show description'}
-              aria-expanded={open}
-              onClick={() => setOpen((v) => !v)}
-            >
-              <svg width="14" height="14" viewBox="0 0 14 14" fill="none" aria-hidden="true">
-                <circle cx="7" cy="7" r="6.5" stroke="currentColor" />
-                <path d="M7 6v4" stroke="currentColor" strokeLinecap="round" strokeWidth="1.25" />
-                <circle cx="7" cy="4.25" r="0.625" fill="currentColor" />
-              </svg>
-            </button>
-            {open && (
-              <div className={styles.infoPopover} role="tooltip">
-                {description}
-              </div>
-            )}
-          </div>
-        )}
+        {description && <FieldHint text={description} />}
       </div>
     </div>
   );
